Change the point shaders so that common code is in shared files.
The shader-loader allows importing chunks into other files.  This allows
common code to be placed in a shared file and imported by other files.
There are some restrictions: the common chunks must have filenames
ending in `glsl`, and the glslangValidator doesn't handle these
inclusions on its own.  However, sharing code is of substantial benefit
for reducing maintenance issues, so it is worth this change.

Note that there are other webpack shader loaders (e.g.,
webpack-glsl-loader) that provide similar include functionality but with
a different syntax, as glsl doesn't have its own include statements.  As
such, if we ever switch the shader loader to another one, the import
statements would have to change.