Add a track feature.
The track feature is a set of lines each of which is composed of points
with position and strictly monotonically increasing time.  There are
utility functions for setting a start and end time.  Two points are
interpolated along the track line: one each for the start and end times.
A marker or text value can be rendered at the end time.

Internally, the track is rendered as three line features; one for all
the points before the start time, one for the start to the end time, and
one for the end time onward.  Additional a marker and/or text feature is
used to render the track heads.